+++ Sources/Ashton/AshtonHTMLWriter.swift
@@ -25,12 +25,17 @@ public final class AshtonHTMLWriter {
     public func encode(_ attributedString: NSAttributedString) -> Ashton.HTML {
         let string = attributedString.string
         let paragraphRanges = self.getParagraphRanges(from: string)
-
+        let fullInputStringLength = (string as NSString).length
         var html = String()
         for paragraphRange in paragraphRanges {
             var paragraphContent = String()
-            let nsParagraphRange = NSRange(paragraphRange, in: string)
+            var nsParagraphRange = NSRange(paragraphRange, in: string)
             var paragraphTag = HTMLTag(defaultName: .p, attributes: [:], ignoreParagraphStyles: false)
+
+            // We add an additional character to the paragraph range to get the parsed paragraph separator (e.g \n)
+            if fullInputStringLength >= nsParagraphRange.upperBound + 1 {
+                nsParagraphRange.length = nsParagraphRange.length + 1
+            }
             // We use `attributedString.string as NSString` casts and NSRange ranges in the block because
             // we experienced outOfBounds crashes when converting NSRange to Range and using it on swift string
             attributedString.enumerateAttributes(in: nsParagraphRange,