+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 ## Unassigned
 
-- [Enhancement] Allow to use Ashton in multithreaded environment
+## 2.2
+
+- [Enhancement] Parse trailing whitespace
 
 ## 2.1
 
+- [Enhancement] Allow to use Ashton in multithreaded environment
 - [Chore] Add additional explicitly dynamic library target
 
 ## 2.0.8